@@ -4,7 +4,7 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "nx": "nx",
     "start": "ng serve",
     "build": "ng build",
@@ -29,56 +29,56 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^11.2.0",
-    "@angular/common": "^11.2.0",
-    "@angular/compiler": "^11.2.0",
-    "@angular/core": "^11.2.0",
-    "@angular/forms": "^11.2.0",
-    "@angular/platform-browser": "^11.2.0",
-    "@angular/platform-browser-dynamic": "^11.2.0",
-    "@angular/router": "^11.2.0",
-    "@nestjs/common": "^7.0.0",
-    "@nestjs/core": "^7.0.0",
-    "@nestjs/platform-express": "^7.0.0",
-    "@nrwl/angular": "12.1.1",
+    "@angular/animations": "14.0.7",
+    "@angular/common": "14.0.7",
+    "@angular/compiler": "14.0.7",
+    "@angular/core": "14.0.7",
+    "@angular/forms": "14.0.7",
+    "@angular/platform-browser": "14.0.7",
+    "@angular/platform-browser-dynamic": "14.0.7",
+    "@angular/router": "14.0.7",
+    "@nestjs/common": "^8.0.0",
+    "@nestjs/core": "^8.0.0",
+    "@nestjs/platform-express": "^8.0.0",
+    "@nrwl/angular": "14.4.3",
     "reflect-metadata": "^0.1.13",
-    "rxjs": "~6.6.3",
+    "rxjs": "^7.0.0",
     "tslib": "^2.0.0",
-    "zone.js": "^0.10.2"
+    "zone.js": "0.11.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1102.0",
-    "@angular-eslint/eslint-plugin": "~4.2.0",
-    "@angular-eslint/eslint-plugin-template": "~4.2.0",
-    "@angular-eslint/template-parser": "~4.2.0",
-    "@angular/cli": "~11.2.0",
-    "@angular/compiler-cli": "^11.2.0",
-    "@angular/language-service": "^11.2.0",
-    "@nestjs/schematics": "^7.0.0",
-    "@nestjs/testing": "^7.0.0",
-    "@nrwl/cli": "12.1.1",
-    "@nrwl/cypress": "12.1.1",
-    "@nrwl/eslint-plugin-nx": "12.1.1",
-    "@nrwl/jest": "12.1.1",
-    "@nrwl/linter": "12.1.1",
-    "@nrwl/nest": "12.1.1",
-    "@nrwl/node": "12.1.1",
-    "@nrwl/tao": "12.1.1",
-    "@nrwl/workspace": "12.1.1",
-    "@types/jest": "26.0.8",
-    "@types/node": "14.14.33",
-    "@typescript-eslint/eslint-plugin": "4.19.0",
-    "@typescript-eslint/parser": "4.19.0",
+    "@angular-devkit/build-angular": "14.0.7",
+    "@angular-eslint/eslint-plugin": "14.0.2",
+    "@angular-eslint/eslint-plugin-template": "14.0.2",
+    "@angular-eslint/template-parser": "14.0.2",
+    "@angular/cli": "~14.0.0",
+    "@angular/compiler-cli": "14.0.7",
+    "@angular/language-service": "14.0.7",
+    "@nestjs/schematics": "^8.0.0",
+    "@nestjs/testing": "^8.0.0",
+    "@nrwl/cli": "14.4.3",
+    "@nrwl/cypress": "14.4.3",
+    "@nrwl/eslint-plugin-nx": "14.4.3",
+    "@nrwl/jest": "14.4.3",
+    "@nrwl/linter": "14.4.3",
+    "@nrwl/nest": "14.4.3",
+    "@nrwl/node": "14.4.3",
+    "@nrwl/workspace": "14.4.3",
+    "@types/jest": "27.4.1",
+    "@types/node": "18.0.0",
+    "@typescript-eslint/eslint-plugin": "5.29.0",
+    "@typescript-eslint/parser": "5.29.0",
     "cypress": "^6.0.1",
-    "dotenv": "8.2.0",
-    "eslint": "7.22.0",
+    "dotenv": "10.0.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "jest": "26.2.2",
-    "jest-preset-angular": "8.4.0",
-    "prettier": "2.2.1",
-    "ts-jest": "26.5.5",
-    "ts-node": "~9.1.1",
-    "typescript": "~4.1.4"
+    "jest": "27.5.1",
+    "jest-preset-angular": "9.0.3",
+    "nx": "14.4.3",
+    "prettier": "2.6.2",
+    "ts-jest": "27.1.4",
+    "ts-node": "10.8.2",
+    "typescript": "4.7.4"
   }
 }
